refactor(services): organize services into single folder
