--- services/bot.go
@@ -1,1 +1,0 @@
-package services
